Change TakeDamage of enemy int to double.

+++ Project/EnemyComponents/EnemyStateGoriya.cs
@@ -134,7 +134,7 @@ namespace Project1.EnemyComponents
         {
             Sprite = SpriteFactory.Instance.GetSpriteData(ID + DirectionState.ID);
         }
-        public void TakeDamage(int damage)
+        public void TakeDamage(double damage)
         {
             Enemy.Health.DecreaseHealth(0 + damage);
         }
